refactor(services): resolve residual subpackage boundary violations (#2862)
Reduce root barrel imports from 16 to 2 documented escape hatches.

Changes:
- Fix shared/branches.py self-reference (resolve_branch_arg now calls itself directly)
- Migrate test imports to subpackage barrels (flow, task, shared, issue)
- Update test patches from "vibe3.services.resolve_branch_arg" to "vibe3.services.shared.branches.resolve_branch_arg"
- Update ROOT_BARREL_IMPORT_BASELINE from 16 to 2

Root barrel imports remaining (documented escape hatches):
1. src/vibe3/services/shared/branches.py:36 - FlowService
2. src/vibe3/services/shared/errors.py:83 - ErrorTrackingService

These temporary bridges cannot be replaced with direct business-subpackage
imports from shared/. Resolution requires proper design change (dependency
injection, moving wrappers, or protocol boundaries).

All modularity tests pass:
- test_services_subpackage_boundaries.py: 10 passed, 2 xfailed (known cycles)
- test_services_reexport_surface.py: all pass
- Root barrel count verified: 2 (as expected)

No production behavior changes. All import changes are in test files or
import-path-only refactors.
