+++ tests/vibe3/execution/test_error_block_decoupling.py
@@ -16,7 +16,7 @@ import pytest
 from vibe3.clients import SQLiteClient
 from vibe3.exceptions import GitHubAPIError
 from vibe3.execution.noop_gate import apply_unified_noop_gate
-from vibe3.services import fail_issue
+from vibe3.services.issue import fail_issue
 
 
 @pytest.fixture
