+++ tests/vibe3/test_modularity/test_services_reexport_surface.py
@@ -197,7 +197,9 @@ KNOWN_PURE_BRIDGES = set()
 
 KNOWN_LAYER_CROSSING_BRIDGES = set()
 
-ROOT_BARREL_IMPORT_BASELINE = 16
+ROOT_BARREL_IMPORT_BASELINE = (
+    2  # #2862: only 2 escape hatches remain (FlowService, ErrorTrackingService)
+)
 SHARED_BARREL_IMPORT_BASELINE = 16
 PURE_BRIDGE_MODULE_BASELINE = len(KNOWN_PURE_BRIDGES)
 LAYER_CROSSING_BRIDGE_BASELINE = len(KNOWN_LAYER_CROSSING_BRIDGES)
@@ -228,12 +230,12 @@ def test_root_barrel_import_count() -> None:
             print(f"     {file}: {count} imports")
 
     assert len(imports) <= ROOT_BARREL_IMPORT_BASELINE, (
-        "Root barrel imports increased beyond the issue #2698 baseline: "
+        "Root barrel imports increased beyond the issue #2862 baseline: "
         f"expected <= {ROOT_BARREL_IMPORT_BASELINE}, found {len(imports)}"
     )
     if imports:
         pytest.xfail(
-            f"Baseline: {len(imports)} root barrel imports remain (issue #2698)"
+            f"Baseline: {len(imports)} root barrel imports remain (issue #2862)"
         )
 
     assert len(imports) == 0, f"Found {len(imports)} root barrel imports"
